Consider number of packets lost between last and before last packets in ControlException
The last packet that is received from Control already sets control_command_success_rate to 0.
If we only show the before last control_command_success_rate in case of exceptions, this can
lead to wrong exceptions messages such as
```
[communication_constraints_violation]
control_command_success_rate: 1
```
This might happen if we had perfect communication and we suddenly loose 20 packets in a row.

To solve this issue, this commit considers also the packets lost between the last and before
last packets in the control_command_succes_rate printed in the exception message.