Define velocity limits depending on accel limits and add factor for Cartesian rotations
Velocity limits are now defined depending on a tolerated number of packets
lost (kTolNumberPacketsLost) and the acceleration limits.
The reason for that is that Control assumes a constant acceleration
model when a packet is lost. With this implementation, velocity limits
will not be violated even if kTolNumberPacketsLost are lost and the robot
is moving at maximum acceleration and velocity.

The Cartesian pose interface has also a factor < 1 for its rotational limits.
Due to the integraton and differentiation of rotational velocities,
slight innacuracies (+/- 5 at a jerk level) could cause a violation of hard
limits. This factor makes sure that this does not occur.